+++ src/routes/(v1)/tests/issue-176/+page.svelte
@@ -20,7 +20,6 @@
 					setError(form, 'email', 'Suspicious email address.');
 				} else if (form.valid) {
 					setMessage(form, 'Valid data!');
-					// TODO: Do something with the validated data
 				}
 			},
 			onError({ result, message }) {